Add ColorParser, ColorExpression and ExpressionsRegistry classes
This is a rough implementation based on current pigments.
The main thing is that managing the expression and using them is now
done in two different entities.
An ExpressionsRegistry instance is exposed as a singleton module that
can be exposed through a service.
A ColorParser uses the expressions registry to transform a string into
a color.
A ColorExpression represents a single expression and can find it and
parse it.